requirements_fixed.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
PyJWT==2.3.0
PyMySQL==1.0.2
redis==4.1.4
loguru==0.6.0
APScheduler==3.9.1
requests==2.27.1
Pillow==9.0.1
